+++ src/types/categories/index.ts
@@ -1,0 +1,7 @@
+export type CategoryData = {
+  id: number;
+  name: string;
+  description: string;
+  color: string;
+  imageUrl: string;
+};
