+++ src/types/order/index.ts
@@ -1,0 +1,15 @@
+export type OrderHistory = {
+  id: number;
+  count: number;
+};
+
+export type OrderFormData = {
+  productId: number;
+  productQuantity: number;
+  messageCardTextMessage: string;
+  senderId: number;
+  receiverId: number;
+  hasCashReceipt: boolean;
+  cashReceiptType?: 'PERSONAL' | 'BUSINESS';
+  cashReceiptNumber?: string;
+};
