+++ src/types/product/index.ts
@@ -1,0 +1,36 @@
+import type { Content, Pageable } from '..';
+
+export type Product = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+};
+
+export type ProductData = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  categoryId: number;
+};
+
+export type ProductOptionsData = {
+  id: number;
+  name: string;
+  quantity: number;
+  productId: number;
+};
+
+export type ProductListResponse = {
+  content: Content[];
+  pageable: Pageable;
+  totalPages: number;
+  totalElements: number;
+  last: boolean;
+  number: number;
+  size: number;
+  numberOfElements: number;
+  first: boolean;
+  empty: boolean;
+};
