+++ src/types/wishlist/index.ts
@@ -1,0 +1,3 @@
+export type WishData = {
+  productId: number;
+};
